Added `ContentManager`, reworked Cycle's, finished Babysteps, wrote Test
1. Gradle
Upgraded build.gradle to work with junit4.12 unit-test

2. ContentManager
Wrote the minimum of the ContentManager to make the test viable

3. CycleEnum (reworked Cycle's)
To make it easier to check in which phase we are and increase the
cohesion I decided to use a enum to represent the phases of the cycle.

4. Babysteps and it's test
The plugin is finished. The test (2 tests in total) work both on my
machine. Let's see what Travis-CI is saying.