InformationCore, PluginManager, Babysteps, Tracker, Settings
Created the InformationCore to add a point where the information will
gather together.

PluginManager has been extended. Waiting for the CompileManager from
Kevin aka Schiggy.

Extended Babysteps, so that it can pull the setting from the
SettingsManager. Later it has to be able to gather this information an
use it.

Created the base for the Tracker.

Settings have been finished.